add explicit casts
git-svn-id: https://svn.r-project.org/R/trunk@64435 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/src/fft.c
@@ -887,7 +887,7 @@ void fft_factor(int n, int *pmaxf, int *pmaxp)
 
     /* extract 3^2, 5^2, ... */
     kchanged = 0;
-    sqrtk = sqrt(k);
+    sqrtk = (int)sqrt(k);
     for (j = 3; j <= sqrtk; j += 2)
     {
         jj = j * j;
@@ -900,7 +900,7 @@ void fft_factor(int n, int *pmaxf, int *pmaxp)
         if (kchanged)
         {
             kchanged = 0;
-            sqrtk = sqrt(k);
+            sqrtk = (int)sqrt(k);
         }
     }
